% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sit_1_2.R
\name{create_example_input}
\alias{create_example_input}
\title{create_example_input}
\usage{
create_example_input()
}
\value{
A list containing the simulation inputs of jim's excel sheet
}
\description{
create_example_input
}
