% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sit_1_2.R
\name{simpop}
\alias{simpop}
\title{Simulate data with mark selective fisheries}
\usage{
simpop(exdata)
}
\arguments{
\item{exdata}{A list containing data inputs, similar to that
produced as an output of create_example_input}
}
\value{
A list containing the simulation outputs of a simple msf fishery
}
\description{
Simulate data with mark selective fisheries
}
