% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sit_1_2.R
\name{sit2}
\alias{sit2}
\title{Simulate and estimate parameters using sit method 2}
\usage{
sit2(simdata)
}
\arguments{
\item{simdata}{A list containing data inputs, similar to that
produced as an output of create_example_input}
}
\value{
A list containing the simulation outputs of a simple msf fishery
}
\description{
Simulate and estimate parameters using sit method 2
}
